Factor out external event notification funcs
This change moves external event notification functionality into
`event-notification.go`. This simplifies notification related code.
